pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
django = "^3.1.3"
djangorestframework = "^3.12.1"
djangorestframework-simplejwt = "^4.4.0"
mysqlclient = "^2.0.1"
django-cors-headers = "^3.5.0"
pillow = "^8.0.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pre-commit = "^2.8.2"
isort = "^5.6.4"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 90
exclude = '''
/(
  | \.git
)/
'''

[tool.isort]
profile = "black"
atomic = true
include_trailing_comma = true
lines_after_imports = 2
use_parentheses = true
src_paths = ["backend"]
filter_files = true
known_third_party = ["django", "rest_framework"]
